Add initial infra setup
Add a Terraform setup that provisions a Kubernetes cluster. That cluster
gets a: Deployment, Service and Ingress so that the container we built
and pushed in an earlier commit is publicly available.

I chose to deploy the Kubernetes config through Terraform, to stick to
one tool. That will likely keep the entire setup a bit simpler. We don't
need any features from Helm or Kustomize for this simple deployment.

To make the authentication for Terraform's Kubernetes provider seamless
(no extra setup required), I split the Terraform config into two
modules: infrastructure and Kubernetes config. That way, the Kubernetes
config can pull in data from the infrastructure and authenticate to the
cluster automagically.