Add image to demo app page
Through Workload Identity, the cluster SAC now has access to the cloud
bucket. Unfortunately, I couldn't figure out how to make Nginx
authenticate itself through Workload Identity to the Cloud Bucket, so I
used a workaround.

The pods download the image from the pod using an init container and
then store it locally for as long as the pod is alive. I did this so
that I could stick to using Nginx. A better alternative might be to
write a small Flask app, which uses a Google library to fetch the image
on-the-fly. The Google libraries/tools seem to properly authenticate
using the correct SAC.

However, I didn't want to write any custom code, so I'll stick to this
approach for now.